Moved Varmint.h to common
